Create OpenAI QA queries - vectorstores
This notebook allows you to upload unlimited text to the OpenAI API, ask questions, and receive extremely high quality answers regarding the uploaded text. This script uses Chroma as a vector store to chunk up the text and retrieve the sentences most likely to contain context and answers about your question. The retrieved sentences, along with the question are uploaded to OpenAI as an embedding. This reduces the token upload amount, which allows for unlimited text to be uploaded and heavily minimizes API fees.